clear compile waring
git-svn-id: http://192.168.20.11/sequoiadb/trunk@22667 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/cls/clsMsgHandler.cpp
@@ -54,10 +54,10 @@ namespace engine
    void _shdMsgHandler::_postMainMsg( const NET_HANDLE & handle,
                                       MsgHeader * pNewMsg )
    {
-      if ( _pShardCB && ( MSG_CAT_NODEGRP_RES == (UINT32)pNewMsg->opCode ||
-           MSG_CAT_QUERY_CATALOG_RSP == (UINT32)pNewMsg->opCode ||
-           MSG_CAT_QUERY_SPACEINFO_RSP == (UINT32)pNewMsg->opCode ||
-           ( MSG_CAT_CATGRP_RES == (UINT32)pNewMsg->opCode &&
+      if ( _pShardCB && ( MSG_CAT_NODEGRP_RES == pNewMsg->opCode ||
+           MSG_CAT_QUERY_CATALOG_RSP == pNewMsg->opCode ||
+           MSG_CAT_QUERY_SPACEINFO_RSP == pNewMsg->opCode ||
+           ( MSG_CAT_CATGRP_RES == pNewMsg->opCode &&
              _pShardCB->getTID() != (UINT32)pNewMsg->requestID ) ) )
       {
          _pShardCB->postEvent( pmdEDUEvent( PMD_EDU_EVENT_MSG,
